requirements.txt
google-generativeai
python-dotenv
streamlit
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image